fix: move Toaster to root layout and remove duplicate ModeWatcher
Fixed toast notifications not appearing in transformation picker window and removed duplicate ModeWatcher.

Changes:
- Moved Toaster from AppLayout to root +layout.svelte
- This ensures toasts appear in all windows, including transformation-picker
- Removed duplicate ModeWatcher from AppLayout (already in root layout)
- Removed duplicate TOASTER_SETTINGS from AppLayout
- Cleaned up unused imports (Toaster, ToasterProps, ModeWatcher, mode)

The transformation picker window now properly shows toast notifications for:
- Loading state: "Running transformation..."
- Error notifications if transformation fails
- Success notifications when complete
